Move translator into api module
